refactor: Extract shared CLUSTER/MERGE toolkit into neutral module
The CLUSTER and MERGE expanders share a set of operator-neutral
expansion helpers -- GenomicColumns, genomic_columns, extract_stranded,
find_projected, require_top_level_projection, reject_cluster_merge_mix,
and transplant -- that lived in cluster.py. That made cluster.py the
de-facto home of cross-operator plumbing and created the package's
first expander-imports-expander edge, so the dependency direction
between the two sibling expanders looked accidental.

Move those helpers into a dedicated neutral module, _genomic.py, so the
shared surface is explicit and each expander stays focused on its own
operator. The module is underscore-prefixed so expander auto-discovery
skips it -- it registers no expander; it is a helper the expanders
import. merge.py now sources the toolkit from the neutral module and
imports only expand_cluster_query from cluster.py, the one deliberate
CLUSTER-reuse composition, since MERGE is built on CLUSTER.

Pure code-move: the helper bodies are byte-identical and the emitted
SQL is unchanged. The full suite stays green.
