+++ src/giql/expanders/cluster.py
@@ -32,7 +32,8 @@ node unchanged so the pass's ``node.replace`` is a no-op — but the *mechanism*
 differs. ``nearest`` does an in-place child ``.replace`` because its LATERAL has a
 parent to replace through; CLUSTER instead copies the enclosing
 :class:`~sqlglot.expressions.Select`, restructures the copy, and *transplants* its
-contents back onto the original SELECT (see :func:`transplant`). That root-
+contents back onto the original SELECT (see
+:func:`giql.expanders._genomic.transplant`). That root-
 preserving transplant is required because the canonical
 ``SELECT *, CLUSTER(...) FROM t`` puts CLUSTER at the *root* ``SELECT``, which has
 no parent to replace it through; transplanting preserves the root's identity (the
@@ -47,38 +48,20 @@ MERGE is built on CLUSTER.
 
 from __future__ import annotations
 
-from typing import NamedTuple
-from typing import TypeVar
-
 from sqlglot import exp
 
-from giql.constants import DEFAULT_CHROM_COL
-from giql.constants import DEFAULT_END_COL
-from giql.constants import DEFAULT_START_COL
-from giql.constants import DEFAULT_STRAND_COL
 from giql.expander import ExpansionContext
 from giql.expander import expand_operators
 from giql.expander import register
+from giql.expanders._genomic import GenomicColumns
+from giql.expanders._genomic import extract_stranded
+from giql.expanders._genomic import find_projected
+from giql.expanders._genomic import genomic_columns
+from giql.expanders._genomic import reject_cluster_merge_mix
+from giql.expanders._genomic import require_top_level_projection
+from giql.expanders._genomic import transplant
 from giql.expressions import GIQLCluster
-from giql.expressions import GIQLMerge
-from giql.table import Tables
 from giql.targets import GenericTarget
-
-_T = TypeVar("_T", bound=exp.Expression)
-
-
-class GenomicColumns(NamedTuple):
-    """The resolved physical column names CLUSTER / MERGE operate over.
-
-    Derived from the enclosing FROM table by :func:`genomic_columns`. A
-    :class:`~typing.NamedTuple` so it still unpacks and indexes positionally
-    (``chrom, start, end, strand = columns``) while giving the four fields names.
-    """
-
-    chrom: str
-    start: str
-    end: str
-    strand: str
 
 
 @register(GenericTarget, GIQLCluster)
@@ -153,57 +136,6 @@ def expand_cluster(node: GIQLCluster, ctx: ExpansionContext) -> exp.Expression:
     return node
 
 
-def genomic_columns(select: exp.Select, tables: Tables) -> GenomicColumns:
-    """Return the ``(chrom, start, end, strand)`` columns for *select*'s FROM table.
-
-    Part of the shared CLUSTER/MERGE expansion toolkit (reused by
-    :mod:`giql.expanders.merge`). Mirrors the legacy
-    ``ClusterTransformer._get_genomic_columns`` / ``_get_table_name``: read the
-    FROM-clause table name, look it up in *tables*, and use its configured column
-    names, falling back to the canonical defaults (and to the default strand
-    column when the table declares none).
-    """
-    table_name: str | None = None
-    from_clause = select.args.get("from_")
-    if from_clause is not None and isinstance(from_clause.this, exp.Table):
-        table_name = from_clause.this.name
-
-    chrom_col = DEFAULT_CHROM_COL
-    start_col = DEFAULT_START_COL
-    end_col = DEFAULT_END_COL
-    strand_col = DEFAULT_STRAND_COL
-
-    if table_name:
-        table = tables.get(table_name)
-        if table:
-            chrom_col = table.chrom_col
-            start_col = table.start_col
-            end_col = table.end_col
-            if table.strand_col:
-                strand_col = table.strand_col
-
-    return GenomicColumns(chrom_col, start_col, end_col, strand_col)
-
-
-def extract_stranded(stranded_expr: exp.Expression | None) -> bool:
-    """Coerce a CLUSTER/MERGE ``stranded`` operand to a bool. Shared toolkit.
-
-    Mirrors the legacy per-transformer coercion exactly: a missing operand is
-    ``False``; an ``exp.Boolean`` yields its raw ``.this``; an ``exp.Literal``
-    compares case-folded to ``TRUE``. The final two arms (``exp.Literal`` and the
-    string-truthiness fallback) are defensive — the GIQL grammar only ever produces
-    ``exp.Boolean`` for ``stranded := <bool>`` — retained for parity with the
-    legacy port.
-    """
-    if stranded_expr is None:
-        return False
-    if isinstance(stranded_expr, exp.Boolean):
-        return stranded_expr.this
-    if isinstance(stranded_expr, exp.Literal):
-        return str(stranded_expr.this).upper() == "TRUE"
-    return str(stranded_expr).upper() in ("TRUE", "1", "YES")
-
-
 def expand_cluster_query(
     query: exp.Select, columns: GenomicColumns
 ) -> exp.Select | None:
@@ -229,72 +161,6 @@ def expand_cluster_query(
     for cluster_expr in cluster_exprs:
         query = _transform_for_cluster(query, cluster_expr, columns)
     return query
-
-
-def find_projected(select: exp.Select, op_type: type[_T]) -> list[_T]:
-    """Return *select*'s projected operators of *op_type* (bare or aliased). Toolkit.
-
-    Shared CLUSTER/MERGE primitive: both expanders and the co-occurrence guard
-    locate their operator the same way — a top-level SELECT projection item that
-    either *is* the operator or is an ``exp.Alias`` wrapping it.
-    """
-    found: list[_T] = []
-    for expression in select.expressions:
-        if isinstance(expression, op_type):
-            found.append(expression)
-        elif isinstance(expression, exp.Alias) and isinstance(expression.this, op_type):
-            found.append(expression.this)
-    return found
-
-
-def require_top_level_projection(
-    select: exp.Select, node: exp.Expression, op_type: type
-) -> None:
-    """Raise if *node* is buried inside a projection expression. Shared toolkit.
-
-    A CLUSTER / MERGE is only expandable as a *top-level* projection item — bare
-    or directly aliased — because the whole-query rewrite restructures the SELECT
-    around it. One nested inside a larger projection expression such as
-    ``ABS(CLUSTER(interval))`` has no coherent rewrite and would otherwise leak an
-    unexpanded operator to the generator, so fail loudly here (#144 A16). An
-    operator that parses *outside* the projection entirely (e.g. in WHERE /
-    ORDER BY) is not under any projection item and is left for the expander's
-    existing no-op path.
-    """
-    operator = op_type.__name__.removeprefix("GIQL").upper()
-    for projection in select.expressions:
-        inner = projection.this if isinstance(projection, exp.Alias) else projection
-        if inner is node:
-            return
-        if any(descendant is node for descendant in inner.walk()):
-            raise ValueError(
-                f"{operator} must be a top-level projection item; it cannot be "
-                "nested inside another expression (e.g. a function call or "
-                "arithmetic)."
-            )
-
-
-def reject_cluster_merge_mix(select: exp.Select) -> None:
-    """Raise if *select* projects both a CLUSTER and a MERGE. Shared toolkit.
-
-    The two are mutually incompatible in one SELECT: MERGE aggregates the rows
-    away (it rewrites the query into a ``GROUP BY`` over a clustered subquery)
-    while CLUSTER is a per-row window over those same rows, so no coherent single
-    query expresses both. The legacy pre-pass chained the two transformers and
-    emitted *non-executable* SQL for this shape (a window over ``GROUP BY``-
-    aggregated rows — a DuckDB ``BinderException``, never a leaked operator). The
-    new in-place ``transplant`` cannot express both at all: whichever expander
-    runs first rewrites the shared SELECT and strands the sibling as an unexpanded
-    node. Fail loudly here — mirroring the ``Multiple MERGE expressions not yet
-    supported`` guard — so the combination raises a clear diagnostic rather than
-    emitting broken SQL.
-    """
-    if find_projected(select, GIQLCluster) and find_projected(select, GIQLMerge):
-        raise ValueError(
-            "CLUSTER and MERGE cannot be combined in a single SELECT; MERGE "
-            "aggregates rows while CLUSTER is a per-row window. Use them in "
-            "separate queries (e.g. CLUSTER over a subquery, or MERGE over one)."
-        )
 
 
 def _transform_for_cluster(
@@ -553,23 +419,3 @@ def _rewrite_predecessor_refs(
     for column in rewritten.find_all(exp.Column):
         column.this.set("quoted", True)
     return rewritten
-
-
-def transplant(select: exp.Select, new: exp.Select) -> None:
-    """Replace *select*'s contents with *new*'s, preserving *select*'s identity.
-
-    Part of the shared CLUSTER/MERGE expansion toolkit. Clears every argument of
-    *select* and re-installs *new*'s, so *select* keeps its position in the
-    surrounding tree (and its identity as the object the pass returns) while taking
-    on the rewritten structure. This is how a whole-query rewrite is applied to a
-    *root* SELECT, which has no parent to ``replace`` through.
-
-    Precondition: *new* MUST be a detached throwaway — a freshly-built
-    ``exp.Select``, as the ``_transform_for_*`` helpers return. Its children are
-    re-parented onto *select*, so passing a node still attached elsewhere would
-    corrupt that other tree.
-    """
-    assert new.parent is None, "transplant() requires a detached `new` subtree"
-    select.args.clear()
-    for key, value in list(new.args.items()):
-        select.set(key, value)
